Fix - puppet and DHCP compete to manage NTP
We want DHCP to configure NTP settings right up until puppet is given the task
of managing a host.  This required the creation of a new network module which
does very little other than what's required for resolving this problem.  It's
pointless to do more with this until such time it is used (if ever) to perform
initial configuration of a host's network settings.