Change - major network module overhaul
This represents a complete rewrite to the network module and implements
a much better approach.

- all host nodes now declare their network interfaces explicitly and
  within in their own class rather than relying on so much site-specific
  detail within the network module
- the network class no longer attempts to manage interfaces; something
  it just barely did before and not very well
- a new network::interface definition provides consistent, detailed
  management of the interfaces
- dropping all support for NetworkManager (other than to remove it) --
  it was an additional factor in an already complex enough job; it could
  certainly be reintroduced at this point, but why?  (I trust systemd
  more than NetworkManager to do the right thing.)

Each host was hand-tested one at a time to ensure no adverse
configuration changes.  Pluto-f15 receives additional enhancements such
that if the network interfaces do get changed, pyptables is notified so
that it can adjust.  (Oddly, this was necessary even though the
configuration was not changing -- apparently it's just sensitive to the
network service being restarted.)